+++ sdk/digitaltwins/azure-resourcemanager-digitaltwins/src/main/java/com/azure/resourcemanager/digitaltwins/AzureDigitalTwinsManager.java
@@ -225,7 +225,7 @@ public final class AzureDigitalTwinsManager {
                 .append("-")
                 .append("com.azure.resourcemanager.digitaltwins")
                 .append("/")
-                .append("1.3.0");
+                .append("1.0.0-beta.1");
             if (!Configuration.getGlobalConfiguration().get("AZURE_TELEMETRY_DISABLED", false)) {
                 userAgentBuilder.append(" (")
                     .append(Configuration.getGlobalConfiguration().get("java.version"))