+++ sdk/digitaltwins/azure-resourcemanager-digitaltwins/CHANGELOG.md
@@ -1,14 +1,29 @@
 # Release History
 
-## 1.4.0-beta.1 (Unreleased)
+## 1.0.0-beta.1 (2024-12-27)
 
-### Features Added
+- Azure Resource Manager AzureDigitalTwins client library for Java. This package contains Microsoft Azure SDK for AzureDigitalTwins Management SDK. Azure Digital Twins Client for managing DigitalTwinsInstance. Package tag package-2025-03-31. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Breaking Changes
 
-### Bugs Fixed
+#### `models.EventGrid` was modified
 
-### Other Changes
+* `provisioningState()` was removed
+* `createdTime()` was removed
+
+#### `models.AzureDataExplorerConnectionProperties` was modified
+
+* `provisioningState()` was removed
+
+#### `models.EventHub` was modified
+
+* `provisioningState()` was removed
+* `createdTime()` was removed
+
+#### `models.ServiceBus` was modified
+
+* `createdTime()` was removed
+* `provisioningState()` was removed
 
 ## 1.3.0 (2024-12-06)